feat(change image function)
This function will select all the carousel buttons and for each of those add an event listener waiting for a click. If the button clicked has as its ID the words next-image then the currentIndex variable value will change adding one and then dividing it by the IMAGES array length. The opposite will happen if another button is clicked. Then display image function is called with the currentIndex value as a parameter.